+++ gradle.properties
@@ -12,7 +12,7 @@ group=org.opentestsystem.ap
 
 version=0.2.185-SNAPSHOT
 
-commonVersion=0.4.215
+commonVersion=0.4.216-SNAPSHOT
 
 configServerVersion=3.1.0.RELEASE
 contentPackageTabulatorVersion=2.2.2
