+++ src/main/java/org/opentestsystem/ap/ivs/IvsConfiguration.java
@@ -1,5 +1,6 @@
 package org.opentestsystem.ap.ivs;
 
+import org.opentestsystem.ap.common.assembler.AppAssembler;
 import org.opentestsystem.ap.common.config.ItemBankProperties;
 import org.opentestsystem.ap.common.datastore.AsyncDataStorePreviewManager;
 import org.opentestsystem.ap.common.datastore.DataStoreAttachmentManager;
@@ -18,9 +19,13 @@ import org.opentestsystem.ap.common.preview.PreviewClient;
 import org.opentestsystem.ap.common.preview.PreviewManager;
 import org.opentestsystem.ap.common.preview.PreviewValidator;
 import org.opentestsystem.ap.common.repository.ItemRepository;
+import org.opentestsystem.ap.common.security.SecurityUtil;
 import org.opentestsystem.ap.common.task.ItemCleanupTask;
 import org.opentestsystem.ap.common.task.TaskProperties;
+import org.opentestsystem.ap.common.validation.IvsProperties;
 import org.opentestsystem.ap.common.validation.ValidationManager;
+import org.opentestsystem.ap.common.validation.ValidationService;
+import org.opentestsystem.ap.common.validation.ValidationUtility;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -152,9 +157,31 @@ public class IvsConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
+    public ValidationUtility validationUtility(IvsProperties ivsProperties) {
+        return new ValidationUtility(   ivsProperties);
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
     public ValidationManager validationManager(DataStorePreviewManager dataStorePreviewManager,
                                                ItemEntityRepository itemEntityRepository,
                                                ItemSyncRepository itemSyncRepository) {
         return new ValidationManager(dataStorePreviewManager, itemEntityRepository, itemSyncRepository);
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    public ValidationService validationService(ValidationManager validationManager,
+                                               IvsProperties ivsProperties,
+                                               SecurityUtil securityUtil,
+                                               ValidationUtility validationUtility,
+                                               AppAssembler appAssembler) {
+        return new ValidationService(
+            validationManager,
+            ivsProperties,
+            securityUtil,
+            validationUtility,
+            appAssembler,
+            this.itemBankProperties);
     }
 }
