--- src/main/java/org/opentestsystem/ap/ivs/IvsProperties.java
@@ -1,29 +1,0 @@
-package org.opentestsystem.ap.ivs;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.stereotype.Component;
-
-import java.util.Arrays;
-import java.util.List;
-
-@Getter
-@Setter
-@Component
-@ConfigurationProperties(prefix = "ivs")
-public class IvsProperties {
-
-    private String cptExecutablePath;
-
-    private String itemBankReportRepoName = "itembankreports";
-
-    private String errorReportFileName = "validation_ErrorReport.csv";
-
-    private List<String> excludedCategories;
-
-    public IvsProperties() {
-        excludedCategories = Arrays.asList("system", "manifest", "exception");
-    }
-}
