+++ src/main/java/org/opentestsystem/ap/ivs/ValidationApi.java
@@ -1,8 +1,8 @@
-package org.opentestsystem.ap.ivs.rest.v1;
+package org.opentestsystem.ap.ivs;
 
 import lombok.extern.slf4j.Slf4j;
-import org.opentestsystem.ap.common.model.ValidationResults;
-import org.opentestsystem.ap.ivs.service.ValidationService;
+import org.opentestsystem.ap.common.model.validation.ValidationResults;
+import org.opentestsystem.ap.common.validation.ValidationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
