--- src/main/java/org/opentestsystem/ap/ivs/models/ErrorReport.java
@@ -1,63 +1,0 @@
-package org.opentestsystem.ap.ivs.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.Data;
-
-import java.util.Objects;
-
-/**
- * A POJO representing the tabulator validation_ErrorReport.csv file.
- */
-@Data
-@JsonPropertyOrder({ "folder", "bankKey", "itemId", "itemType", "category", "severity", "errorMessage", "detail" })
-public class ErrorReport {
-
-    @JsonProperty("Folder")
-    private String folder;
-
-    @JsonProperty("BankKey")
-    private String bankKey;
-
-    @JsonProperty("ItemId")
-    private String itemId;
-
-    @JsonProperty("ItemType")
-    private String itemType;
-
-    @JsonProperty("Category")
-    private String category;
-
-    @JsonProperty("Severity")
-    private String severity;
-
-    @JsonProperty("ErrorMessage")
-    private String errorMessage;
-
-    @JsonProperty("Detail")
-    private String detail;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        ErrorReport that = (ErrorReport) o;
-        return Objects.equals(folder, that.folder) &&
-            Objects.equals(bankKey, that.bankKey) &&
-            Objects.equals(itemId, that.itemId) &&
-            Objects.equals(itemType, that.itemType) &&
-            Objects.equals(category, that.category) &&
-            Objects.equals(severity, that.severity) &&
-            Objects.equals(errorMessage, that.errorMessage) &&
-            Objects.equals(detail, that.detail);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(folder, bankKey, itemId, itemType, category, severity, errorMessage, detail);
-    }
-}
