+++ src/test/java/org/opentestsystem/ap/ivs/TestHelper.java
@@ -4,15 +4,15 @@ import freemarker.template.Configuration;
 import freemarker.template.TemplateExceptionHandler;
 import lombok.Getter;
 import org.opentestsystem.ap.common.assembler.AppAssembler;
-import org.opentestsystem.ap.common.saaif.StringAssembler;
 import org.opentestsystem.ap.common.config.ItemBankProperties;
 import org.opentestsystem.ap.common.itembank.IdGenerator;
 import org.opentestsystem.ap.common.model.JsonModelAssembler;
 import org.opentestsystem.ap.common.saaif.SaaifAssembler;
 import org.opentestsystem.ap.common.saaif.SaaifMetadataAssembler;
 import org.opentestsystem.ap.common.saaif.SaaifWordListAssembler;
-import org.opentestsystem.ap.common.validation.ValidationResultAssembler;
+import org.opentestsystem.ap.common.saaif.StringAssembler;
 import org.opentestsystem.ap.common.util.RandomIdGenerator;
+import org.opentestsystem.ap.common.validation.IvsProperties;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
@@ -65,8 +65,6 @@ public class TestHelper {
 
     private JsonModelAssembler jsonModelAssembler;
 
-    private ValidationResultAssembler validationResultAssembler;
-
     private StringAssembler stringAssembler;
 
     private Configuration freemarker;
@@ -113,11 +111,9 @@ public class TestHelper {
 
         jsonModelAssembler = new JsonModelAssembler();
 
-        validationResultAssembler = new ValidationResultAssembler();
-
         stringAssembler = new StringAssembler(freemarker);
 
-        assembler = new AppAssembler(saaifAssembler, jsonModelAssembler, validationResultAssembler, stringAssembler);
+        assembler = new AppAssembler(saaifAssembler, jsonModelAssembler, stringAssembler);
     }
 
     public void unzipFile(final File zipFile, final Path unzipDestination) throws IOException {
