+++ src/test/java/org/opentestsystem/ap/ivs/config/AppTestConfig.java
@@ -30,8 +30,8 @@ import org.opentestsystem.ap.common.preview.PreviewClient;
 import org.opentestsystem.ap.common.preview.PreviewManager;
 import org.opentestsystem.ap.common.preview.PreviewValidator;
 import org.opentestsystem.ap.common.validation.ValidationManager;
+import org.opentestsystem.ap.common.validation.ValidationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -73,7 +73,6 @@ public class AppTestConfig {
             setGroup(mock(Namespace.class));
         }
     }
-
 
     @Bean
     public AmazonClient amazonClient() {
@@ -129,4 +128,9 @@ public class AppTestConfig {
     public ValidationManager validationManager() {
         return mock(ValidationManager.class);
     }
+
+//    @Bean
+//    public ValidationService validationSerivce() {
+//        return mock(ValidationService.class);
+//    }
 }
