+++ src/test/java/org/opentestsystem/ap/ivs/rest/v1/ValidationApiTest.java
@@ -6,8 +6,9 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
-import org.opentestsystem.ap.common.model.ValidationResults;
-import org.opentestsystem.ap.ivs.service.ValidationService;
+import org.opentestsystem.ap.common.model.validation.ValidationResults;
+import org.opentestsystem.ap.common.validation.ValidationService;
+import org.opentestsystem.ap.ivs.ValidationApi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
