+++ src/test/java/org/opentestsystem/ap/ivs/service/ValidationUtilityTest.java
@@ -21,11 +21,13 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.opentestsystem.ap.common.assembler.AppAssembler;
-import org.opentestsystem.ap.common.model.ValidationResult;
+import org.opentestsystem.ap.common.model.validation.ErrorReport;
+import org.opentestsystem.ap.common.model.validation.ValidationResult;
 import org.opentestsystem.ap.common.saaif.ItemContext;
 import org.opentestsystem.ap.common.validation.ValidationManager;
+import org.opentestsystem.ap.common.validation.ValidationService;
+import org.opentestsystem.ap.common.validation.ValidationUtility;
 import org.opentestsystem.ap.ivs.TestHelper;
-import org.opentestsystem.ap.ivs.models.ErrorReport;
 
 import java.io.IOException;
 import java.net.URL;
